[Refactoring] Move FindRenameRangesAnnotatingConsumer to its own file
